[💄style]: 테마 수정에 따른 코드 포맷팅
